+++ server/src-lib/Hasura/Backends/MSSQL/Instances/Execute.hs
@@ -226,7 +226,7 @@ msDBMutationPlan ::
 msDBMutationPlan userInfo stringifyNum sourceName sourceConfig mrf = do
   go <$> case mrf of
     MDBInsert annInsert -> executeInsert userInfo stringifyNum sourceConfig annInsert
-    MDBDelete _annDelete -> throw400 NotSupported "delete mutations are not supported in MSSQL"
+    MDBDelete annDelete -> executeDelete userInfo stringifyNum sourceConfig annDelete
     MDBFunction {} -> throw400 NotSupported "function mutations are not supported in MSSQL"
   where
     go v = DBStepInfo @'MSSQL sourceName sourceConfig Nothing v
@@ -313,7 +313,7 @@ executeInsert userInfo stringifyNum sourceConfig annInsert = do
         Tx.unitQueryE fromMSSQLTxError $
           toQueryFlat $
             TQ.fromSetIdentityInsert $
-              SetIdenityInsert (_aiTableName $ _aiData insert) SetON
+              SetIdentityInsert (_aiTableName $ _aiData insert) SetON
 
       -- Generate the INSERT query
       let insertQuery = toQueryFlat $ TQ.fromInsert $ TSQL.fromInsert insert
@@ -383,6 +383,62 @@ executeInsert userInfo stringifyNum sourceConfig annInsert = do
             ExpressionProjection $ Aliased (SelectExpression (generateCheckConstraintSelect checkBoolExp)) "check_constraint_select"
        in emptySelect {selectProjections = [mutationOutputProjection, checkConstraintProjection]}
 
+-- Delete
+
+-- | Executes a Delete IR AST and return results as JSON.
+executeDelete ::
+  MonadError QErr m =>
+  UserInfo ->
+  Bool ->
+  SourceConfig 'MSSQL ->
+  AnnDelG 'MSSQL (Const Void) (UnpreparedValue 'MSSQL) ->
+  m (ExceptT QErr IO EncJSON)
+executeDelete userInfo stringifyNum sourceConfig deleteOperation = do
+  preparedDelete <- traverse (prepareValueQuery $ _uiSession userInfo) deleteOperation
+  let pool = _mscConnectionPool sourceConfig
+  pure $ withMSSQLPool pool $ Tx.runTxE fromMSSQLTxError (buildDeleteTx preparedDelete stringifyNum)
+
+-- | Converts a Delete IR AST to a transaction of three delete sql statements.
+--
+-- A GraphQL delete mutation does two things:
+--
+-- 1. Deletes rows in a table according to some predicate
+-- 2. (Potentially) returns the deleted rows (including relationships) as JSON
+--
+-- In order to complete these 2 things we need 3 SQL statements:
+--
+-- 1. SELECT INTO <temp_table> WHERE <false> - creates a temporary table
+--    with the same schema as the original table in which we'll store the deleted rows
+--    from the table we are deleting
+-- 2. DELETE FROM with OUTPUT - deletes the rows from the table and inserts the
+--   deleted rows to the temporary table from (1)
+-- 3. SELECT - constructs the @returning@ query from the temporary table, including
+--   relationships with other tables.
+buildDeleteTx ::
+  AnnDel 'MSSQL ->
+  Bool ->
+  Tx.TxET QErr IO EncJSON
+buildDeleteTx deleteOperation stringifyNum = do
+  let tempTableName = TSQL.TempTableName "deleted"
+      withAlias = "with_alias"
+      createTempTableQuery = toQueryFlat $ TQ.fromSelectIntoTempTable $ TSQL.toSelectIntoTempTable tempTableName deleteOperation
+  -- Create a temp table
+  Tx.unitQueryE fromMSSQLTxError createTempTableQuery
+  let deleteQuery = TQ.fromDelete tempTableName <$> TSQL.fromDelete deleteOperation
+  deleteQueryValidated <- toQueryFlat <$> V.runValidate (runFromIr deleteQuery) `onLeft` (throw500 . tshow)
+  -- Execute DELETE statement
+  Tx.unitQueryE fromMSSQLTxError deleteQueryValidated
+  mutationOutputSelect <- mkMutationOutputSelect stringifyNum withAlias $ dqp1Output deleteOperation
+  let withSelect =
+        emptySelect
+          { selectProjections = [StarProjection],
+            selectFrom = Just $ FromTempTable $ Aliased tempTableName "deleted_alias"
+          }
+      finalMutationOutputSelect = mutationOutputSelect {selectWith = Just $ With $ pure $ Aliased withSelect withAlias}
+      mutationOutputSelectQuery = toQueryFlat $ TQ.fromSelect finalMutationOutputSelect
+  -- Execute SELECT query and fetch mutation response
+  encJFromText <$> Tx.singleRowQueryE fromMSSQLTxError mutationOutputSelectQuery
+
 -- | Generate a SQL SELECT statement which outputs the mutation response
 --
 -- For multi row inserts:
@@ -400,28 +456,38 @@ mkMutationOutputSelect stringifyNum withAlias = \case
     projections <- forM multiRowFields $ \(fieldName, field') -> do
       let mkProjection = ExpressionProjection . flip Aliased (getFieldNameTxt fieldName) . SelectExpression
       mkProjection <$> case field' of
-        IR.MCount -> pure countSelect
+        IR.MCount -> pure $ countSelect withAlias
         IR.MExp t -> pure $ textSelect t
-        IR.MRet returningFields -> mkSelect JASMultipleRows returningFields
+        IR.MRet returningFields -> mkSelect stringifyNum withAlias JASMultipleRows returningFields
     let forJson = JsonFor $ ForJson JsonSingleton NoRoot
     pure emptySelect {selectFor = forJson, selectProjections = projections}
-  IR.MOutSinglerowObject singleRowField -> mkSelect JASSingleObject singleRowField
-  where
-    mkSelect jsonAggSelect annFields = do
-      let annSelect = IR.AnnSelectG annFields (IR.FromIdentifier withAlias) IR.noTablePermissions IR.noSelectArgs stringifyNum
-      V.runValidate (runFromIr $ mkSQLSelect jsonAggSelect annSelect) `onLeft` (throw500 . tshow)
+  IR.MOutSinglerowObject singleRowField -> mkSelect stringifyNum withAlias JASSingleObject singleRowField
 
-    -- SELECT COUNT(*) FROM [with_alias]
-    countSelect =
-      let countProjection = AggregateProjection $ Aliased (CountAggregate StarCountable) "count"
-       in emptySelect
-            { selectProjections = [countProjection],
-              selectFrom = Just $ TSQL.FromIdentifier withAlias
-            }
+mkSelect ::
+  MonadError QErr m =>
+  Bool ->
+  Text ->
+  JsonAggSelect ->
+  Fields (AnnFieldG 'MSSQL (Const Void) Expression) ->
+  m Select
+mkSelect stringifyNum withAlias jsonAggSelect annFields = do
+  let annSelect = IR.AnnSelectG annFields (IR.FromIdentifier withAlias) IR.noTablePermissions IR.noSelectArgs stringifyNum
+  V.runValidate (runFromIr $ mkSQLSelect jsonAggSelect annSelect) `onLeft` (throw500 . tshow)
 
-    textSelect t =
-      let textProjection = ExpressionProjection $ Aliased (ValueExpression (ODBC.TextValue t)) "exp"
-       in emptySelect {selectProjections = [textProjection]}
+-- SELECT COUNT(*) AS "count" FROM [with_alias]
+countSelect :: Text -> Select
+countSelect withAlias =
+  let countProjection = AggregateProjection $ Aliased (CountAggregate StarCountable) "count"
+   in emptySelect
+        { selectProjections = [countProjection],
+          selectFrom = Just $ TSQL.FromIdentifier withAlias
+        }
+
+-- SELECT '<text-value>' AS "exp"
+textSelect :: Text -> Select
+textSelect t =
+  let textProjection = ExpressionProjection $ Aliased (ValueExpression (ODBC.TextValue t)) "exp"
+   in emptySelect {selectProjections = [textProjection]}
 
 -- subscription